1，修改订单统计页面 2，修改业务报表页面

+++ app/Http/Controllers/Api/V1/MyGoodsController.php
@@ -135,7 +135,7 @@ class MyGoodsController extends Controller
             }
 
             if ($goods->fill($attributes)->save()) {
-                $goods->goodsPieces->delete();
+                $goods->goodsPieces && $goods->goodsPieces->delete();
                 //更新商品单位
                 $piecesAttributes = $request->only('pieces_level_1', 'pieces_level_2', 'pieces_level_3', 'system_1',
                     'system_2', 'specification');
